+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-07-30 21:12:56
-// Commit: fix(ui): fix memory leak by cancelling ui subscription
+// Updated at: 2026-08-03 14:19:27
+// Commit: test(ui): add unit tests for ui business logic
 
 class uiComponentState {
-  final String commitRef = '3265';
-  final String lastMessage = 'fix(ui): fix memory leak by cancelling ui subscription';
-  final DateTime timestamp = DateTime.parse('2026-07-30T21:12:56+05:00');
+  final String commitRef = '3302';
+  final String lastMessage = 'test(ui): add unit tests for ui business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-03T14:19:27+05:00');
 }